Further iterated output_verification.py using Claude
Trying to improve ergonomics before we add many more tests.

Prompt:

Hmm... running the test this many times is quite tedious. Can we modify the output_verification even more so that it creates all references it can in one go?

And I think we can modify the behavior to pass if the reference file is created (after all it must match in that case). Since the reference files are in git, it is the developers responsibility to compare the reference files before committing them if they are changed (and we only change them if they are deleted, so it can't be complete accident.)

We could log some warnings when references are created, but still have the output_verification go to the end and ignore files for which it creates a reference.

---