Claude integrated the new stages idea to plan.md
prompt:

---

Hmm.. good. Now to thanks to the file links you added I see that we'll actually want to add a new "stages" as part of this plan.

Please check the temporary section I added to plan.md and modify the numbered steps below it to mention this idea.

We could create these new stages as separate files and keep the old ones until we are ready to replace the old ones.