Claude added chat merging support + test
It made one mistake and tried to remove the asserts from the
existing test and tried to compare files by name only (not id).
Second prompt I interjected fixed that.

prompt:
To save context space and keep the chat compact:

- Only do the file changes
- Don't describe before or after what is being done

Other instructions:
- Please don't apologize for typos, import issues etc, just fix and
  move on. IDE tells me what is going on.
- After every major change, either of the following to verify tests, linter etc are ok:
  - ```./build.sh```
  - ```./build --refresh-test-resources``` (if we expect test output changes)

Please study the provided files for context.

We will resolve the TODO in the middle of message_processor.

- We will add parameter old_chat_data: ChatData
- Zip file support will be added later and VFS will abstract it, so it is not concern here
- The new message processor will take on the concerns of the merge_chat_data
- message processor can assume the VFS it receives contains both new from input dir and old information from the output dir
- however, it must accept the old ChatData coming from JSON in the output folder.
- it must first check the old chat data and collect chats with file ids that either exist according to vfs or don't
- for file ids present in the old data we must not parse again even if the file still exists, we can just use the Chat structure from the old data
- only the new chat file ids must be processed as is done now in the code
- however the chat files must be ordered by their modification time stamp even if the actual file does not exist, we can get the info from vfs

---

We must be careful with the check you added here. It should be based on ChatFileID which will detect if there is a different file with the same name.

---

Lets add a test case for the new functionality