Claude fixing bugs in output_file handling
Still not perfect, but much better

prompt:

Nice work! Now 2022 html is generated, but we have more details to fix.

We can see that the chat_dependencies, css_dependency and media_dependencies are not populated for the second output file, which probably leads to the generate flag remaining false.

We should look at why the dependencies are not populated.

I can see that html_generator creates OutputFile instances and media_locator as well, I don't think that makes sense. Those should be removed.

I think we need one more stage, lets put that too in a separate .py and integrate in main.py

The new stage will scan through Chats in ChatData and generate a fresh set of output_files.

Then the check_output_dependencies will mark which output files need to be regenerated.