Claude helped fix output instability caused by css mtime
prompt:
---
We need to stabilise the timestamp of the browseability-generator.css.

Since the html_generator.py and output_file_generator.py get its location from the source file, this causes an issue. We don't want to go poking mtime of a git managed file.

We should change these places to instead go from the current working directory and basically assume it is the workspace root.

I think also it would be better if the browseability-generator.css was in its own folder say src/resources or src/css or just css.

And perhaps we should have a small file for handling it. Say css.py or something.
---
Please modify test_integration.py so that it copies the css to a subdir the temp directory and force its mtime. It can then change cwd to the temp dir to make the program pick up the css with fixed mtime.
---
Please run this instead:
./build.sh --refresh-test-resources
---
I think we both forgot that we were supposed to remove the css_setup.py. I just deleted it.

Now please run ./build.sh --refresh-test-resources