Claude implemented logging
I tested the program with a large cache of production data. It worked surprisingly well, but it didn't find any chats in .zip files, but it did find chats in a recent zip file I had manually unpacked.

Also it runs for some time and there is no feedback, so it is hard to know what is going on.

I think we need 4 level verbose modes. Default is 1 with minimal output at ket points in the main program logging the number of seen messages, files etc. 0 is no output.

2 outputs data during vfs scanning every second giving the number of found _chat files and zip files and the directory currently being scanned.

Level 4 logs every file seen and later we can also add verbose debug logs to other places in the program when level 4 is on.

Maybe internally we'd call these, but for user it is just verbosity level:
FATAL
INFO
DEBUG
TRACE

Lets first add the logging system first and worry about the zip file issue later.

I believe I can use logging to quickly find out the issue with the larger set of test data.